+++ Sources/SGCore/Threading/ThreadsManager.h
@@ -7,6 +7,7 @@
 
 #include <SGCore/pch.h>
 
+#include "WrappedObject.h"
 #include "Thread.h"
 #include "ThreadsPool.h"
 
@@ -17,7 +18,7 @@ namespace SGCore::Threading
         friend struct Thread;
         
         static std::shared_ptr<Thread> currentThread() noexcept;
-        
+
         static std::shared_ptr<Thread> getMainThread() noexcept;
 
         SG_NOINLINE static ThreadsPool& getGlobalPool() noexcept;
@@ -25,9 +26,7 @@ namespace SGCore::Threading
     private:
         static inline ThreadsPool m_globalThreadsPool;
         
-        static inline std::mutex m_threadAccessMutex;
-        
-        static inline std::vector<std::shared_ptr<Thread>> m_threads;
+        static inline WrappedObject<std::vector<std::shared_ptr<Thread>>> m_threads;
         
         static inline std::shared_ptr<MainThread> m_mainThread;
 
@@ -35,7 +34,7 @@ namespace SGCore::Threading
             m_mainThread = std::shared_ptr<MainThread>(new MainThread);
             m_mainThread->m_nativeThreadID = std::this_thread::get_id();
             m_mainThread->m_sleepIfNotBusy = false;
-            m_threads.push_back(m_mainThread);
+            m_threads.getWrapped().push_back(m_mainThread);
             
             return true;
         }();
